requirements/ml.txt
# ML and NLP dependencies - Complete list
torch==2.1.0
transformers==4.35.2
tokenizers==0.15.0
accelerate==0.24.1
datasets==2.14.6
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
huggingface-hub>=0.16.4,<1.0

# Vector search and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Experiment tracking
wandb==0.15.5

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.0
python-dotenv==1.0.0
Jinja2==3.1.2